build: use rollup
