+++ lib/transform.js
@@ -1,11 +1,9 @@
 import crypto from 'crypto';
-import fs from 'fs';
-import path from 'path';
 import * as datetime from './datetime.js';
 
-const usStates = JSON.parse(fs.readFileSync(path.join('coronavirus-data-sources', 'lib', 'us-states.json'), 'utf8'));
+import usStates from '../coronavirus-data-sources/lib/us-states.json';
 
-const countryCodes = JSON.parse(fs.readFileSync(path.join('coronavirus-data-sources', 'ISO-3166-Countries-with-Regional-Codes', 'slim-3', 'slim-3.json'), 'utf8'));
+import countryCodes from '../coronavirus-data-sources/ISO-3166-Countries-with-Regional-Codes/slim-3/slim-3.json';
 
 /*
   Override some incorrect country names
