+++ package.json
@@ -2,7 +2,7 @@
   "name": "coronadatascraper",
   "version": "1.0.0",
   "description": "A scraper for Coronavirus data",
-  "main": "cli.js",
+  "main": "index.js",
   "author": "Larry Davis <lazdnet@gmail.com>",
   "license": "BSD",
   "type": "commonjs",
@@ -10,10 +10,9 @@
     "node": ">=12.15.0 <13.0.0"
   },
   "scripts": {
-    "dev": "yarn buildSite & watch 'yarn buildSite' site & yarn serveSite",
-    "serveSite": "npx npx live-server dist",
+    "dev": "npx rollup -c --watch",
     "build": "yarn start && yarn timeseries && yarn buildSite",
-    "buildSite": "mkdir -p dist/; cp -r site/* dist/ && ./tools/copySpectrumCSS.sh",
+    "buildSite": "npx rollup -c",
     "dateData": "DATE=$(date +'%Y-%m-%d') && cp dist/data.json dist/data-$DATE.json && cp dist/features.json dist/features-$DATE.json && cp dist/data.csv dist/data-$DATE.csv",
     "deploy": "npm run dateData && npm run deploySite",
     "deploySite": "npm run buildSite && gh-pages -a -d dist/ -e .",
@@ -27,6 +26,7 @@
   "dependencies": {
     "@adobe/leonardo-contrast-colors": "^1.0.0-alpha.4",
     "@adobe/spectrum-css": "^2.18.0",
+    "@rollup/plugin-json": "^4.0.2",
     "@turf/turf": "^5.1.6",
     "cheerio": "^1.0.0-rc.3",
     "csv-parse": "^4.8.8",
@@ -37,12 +37,13 @@
     "fast-glob": "^3.2.2",
     "gh-pages": "^2.2.0",
     "jest-each": "^25.1.0",
-    "live-server": "^1.2.1",
     "mock-fs": "^4.11.0",
     "needle": "^2.3.3",
+    "postcss-nested": "^4.2.1",
     "puppeteer": "^2.1.1",
     "request": "^2.88.2",
-    "watch": "^1.0.2",
+    "rollup": "^2.1.0",
+    "rollup-plugin-serve": "^1.0.1",
     "yargs": "^15.3.1"
   },
   "devDependencies": {
@@ -59,7 +60,10 @@
     "husky": "^4.2.3",
     "jest": "^25.1.0",
     "lint-staged": "^10.0.8",
-    "prettier": "^1.19.1"
+    "postcss-import": "^12.0.1",
+    "prettier": "^1.19.1",
+    "rollup-plugin-copy": "^3.3.0",
+    "rollup-plugin-postcss": "^2.5.0"
   },
   "husky": {
     "hooks": {
