+++ site/icons/style.css
@@ -1,10 +1,10 @@
 @font-face {
   font-family: 'icomoon';
-  src:  url('fonts/icomoon.eot?s1jly5');
-  src:  url('fonts/icomoon.eot?s1jly5#iefix') format('embedded-opentype'),
-    url('fonts/icomoon.ttf?s1jly5') format('truetype'),
-    url('fonts/icomoon.woff?s1jly5') format('woff'),
-    url('fonts/icomoon.svg?s1jly5#icomoon') format('svg');
+  src:  url('./icons/fonts/icomoon.eot?s1jly5');
+  src:  url('./icons/fonts/icomoon.eot?s1jly5#iefix') format('embedded-opentype'),
+    url('./icons/fonts/icomoon.ttf?s1jly5') format('truetype'),
+    url('./icons/fonts/icomoon.woff?s1jly5') format('woff'),
+    url('./icons/fonts/icomoon.svg?s1jly5#icomoon') format('svg');
   font-weight: normal;
   font-style: normal;
   font-display: block;
