+++ site/index.html
@@ -12,11 +12,6 @@
   <meta name="twitter:description" content="COVID-19 Coronavirus data scraped from government and curated data sources.">
   <meta name="viewport" content="initial-scale=1">
 
-  <link rel="stylesheet" type="text/css" href="spectrum-css.css">
-  <link rel="stylesheet" type="text/css" href="colors.css">
-  <link rel="stylesheet" type="text/css" href="style.css">
-  <link rel="stylesheet" type="text/css" href="icons/style.css">
-
   <script src="https://cdn.jsdelivr.net/npm/handsontable/dist/handsontable.full.min.js"></script>
   <link href="https://cdn.jsdelivr.net/npm/handsontable/dist/handsontable.full.min.css" rel="stylesheet">
   <script src="https://cdnjs.cloudflare.com/ajax/libs/PapaParse/5.1.0/papaparse.min.js"></script>
